Restructured the NavHeader to now be the parent of the NavDrawer, this also allowed signout logic to move to the drawer itself and off of each page. This will allow the NavHeader and drawer to seamlessly integrate with whatever page they are rendered in.
