+++ timekeeper/src/scenes/SettingsPage/SettingsPage.tsx
@@ -1,0 +1,7 @@
+import { Box } from "@mui/material";
+
+const SettingsPage = () => {
+  return <Box></Box>;
+};
+
+export default SettingsPage;
